# source/frontend/src/app/core/api/generated/openapi-core.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/arquivos": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Optional field name for sorting.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive text filter for representative read endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paged document attachment metadata."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/arquivos/presigned-upload": {
      "post": {
        "operationId": "DocumentsController_presignUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignUploadRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a presigned upload URL for S3."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/arquivos/{anexo_id}/confirmar": {
      "patch": {
        "operationId": "DocumentsController_registerByPath",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Register a completed upload as attachment metadata."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/arquivos/{id}": {
      "delete": {
        "operationId": "DocumentsController_deleteAttachment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a document attachment metadata record."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/arquivos/{id}/download": {
      "get": {
        "operationId": "DocumentsController_presignDownload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Create a short-lived presigned download URL."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "SGP API",
    "description": "Modern SGP NestJS API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "PresignUploadRequestDto": {
        "type": "object",
        "properties": {
          "ownerType": {
            "type": "string",
            "maxLength": 120
          },
          "ownerId": {
            "type": "string",
            "maxLength": 120
          },
          "fileName": {
            "type": "string",
            "maxLength": 255
          },
          "contentType": {
            "type": "string",
            "maxLength": 255
          },
          "sizeBytes": {
            "type": "number",
            "minimum": 1,
            "maximum": 1073741824
          }
        },
        "required": [
          "ownerType",
          "fileName",
          "contentType"
        ]
      },
...
    }
  }
}